+++ Release_Notes.md
@@ -135,7 +135,9 @@
 - DSearch Added a constructor that accepts connectionString and the securityEntitySeqID so we don't have to set the properties in BSearch.
     - BSearch changed the constructor to pass the connectionString and the securityEntitySeqID to the DSearch constructor
     - BSearch now accounts for CentralManagement
-
+- DSecurityEntities Added a constructor that accepts connectionString so we don't have to set the properties in BSecurityEntities.
+    - BSecurityEntities changed the constructor to pass the connectionString to the DSecurityEntities constructor
+    - BSecurityEntities now accounts for CentralManagement
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure